+++ src/main/java/org/octri/omop_annotator/view/MeasurementRow.java
@@ -8,9 +8,9 @@ import java.util.Date;
  */
 public interface MeasurementRow {
 
-	public Long getId();
+	public Integer getId();
 
-	public Long getPerson();
+	public Integer getPerson();
 
 	public Date getMeasurementDatetime();
 
@@ -24,6 +24,6 @@ public interface MeasurementRow {
 
 	public String getUnit();
 
-	public Long getVisitOccurrence();
+	public Integer getVisitOccurrence();
 
 }
